correct id note

+++ app/src/main/java/com/angelomoroni/kotlintexteditor/models/Note.kt
@@ -11,8 +11,8 @@ data class Note (var title: String,
                  var body: String = "",
                  var creatTime: Long = System.currentTimeMillis(),
                  var lastModTime: Long = System.currentTimeMillis(),
-                 var id: Int = -1) : Parcelable {
-    constructor(source: Parcel): this(source.readString(), source.readString(), source.readLong(), source.readLong(), source.readInt())
+                 var id: Long = -1) : Parcelable {
+    constructor(source: Parcel): this(source.readString(), source.readString(), source.readLong(), source.readLong(), source.readLong())
 
     override fun describeContents(): Int {
         return 0
@@ -23,7 +23,7 @@ data class Note (var title: String,
         dest?.writeString(body)
         dest?.writeLong(creatTime)
         dest?.writeLong(lastModTime)
-        dest?.writeInt(id)
+        dest?.writeLong(id)
     }
 
     companion object {
